% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{add_collection_alias_store}
\alias{add_collection_alias_store}
\title{Add a collection alias}
\usage{
add_collection_alias_store(store_ptr, namespace, alias, digest)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{namespace}{Alias namespace}

\item{alias}{Alias name}

\item{digest}{Collection digest}
}
\description{
Add a collection alias
}
