% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{add_fasta_store}
\alias{add_fasta_store}
\title{Add a FASTA file with force option}
\usage{
add_fasta_store(store_ptr, file_path, force)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{file_path}{Path to FASTA file}

\item{force}{Whether to overwrite existing sequences/collections}
}
\description{
Add a FASTA file with force option
}
