% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{add_sequence_alias_store}
\alias{add_sequence_alias_store}
\title{Add a sequence alias}
\usage{
add_sequence_alias_store(store_ptr, namespace, alias, digest)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{namespace}{Alias namespace}

\item{alias}{Alias name}

\item{digest}{Sequence digest}
}
\description{
Add a sequence alias
}
