% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-class.R
\name{as_granges}
\alias{as_granges}
\title{Convert RegionSet to GRanges}
\usage{
as_granges(rs)
}
\arguments{
\item{rs}{A RegionSet object}
}
\value{
A GRanges object
}
\description{
Converts a RegionSet to a GRanges object. Coordinates are
converted from 0-based half-open (BED) to 1-based closed (GRanges).
Strand information is preserved.
}
