% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-class.R
\name{as_regionset}
\alias{as_regionset}
\title{Convert to RegionSet}
\usage{
as_regionset(x)
}
\arguments{
\item{x}{Input to convert (see \code{\link{RegionSet}} for accepted types)}
}
\value{
A \code{RegionSet} object
}
\description{
Converts a GRanges object, file path, data.frame, or
externalptr to a RegionSet. Alias for \code{\link{RegionSet}}.
}
