% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{build_restricted_universe}
\alias{build_restricted_universe}
\title{Build a restricted universe from user sets.}
\usage{
build_restricted_universe(user_sets_list)
}
\arguments{
\item{user_sets_list}{R list of RegionSet external pointers}
}
\description{
Build a restricted universe from user sets.
}
