% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lola.R
\name{buildRestrictedUniverse}
\alias{buildRestrictedUniverse}
\title{Build a restricted universe from user sets}
\usage{
buildRestrictedUniverse(userSets)
}
\arguments{
\item{userSets}{A list of RegionSets}
}
\value{
A RegionSet representing the restricted universe
}
\description{
Build a restricted universe from user sets
}
