% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicdist.R
\name{calcTSSDist}
\alias{calcTSSDist}
\title{Calculate TSS distances (absolute)}
\usage{
calcTSSDist(query, features)
}
\arguments{
\item{query}{A GRanges object, file path, data.frame, or RegionSet}

\item{features}{A GRanges object, file path, data.frame, or RegionSet}
}
\value{
Integer vector of absolute distances
}
\description{
Calculates the absolute distance from each query region midpoint
to the nearest feature midpoint.
}
