% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{check_universe}
\alias{check_universe}
\title{Check universe appropriateness for user sets.}
\usage{
check_universe(user_sets_list, universe_ptr)
}
\arguments{
\item{user_sets_list}{R list of RegionSet external pointers}

\item{universe_ptr}{ExternalPtr to universe RegionSet}
}
\description{
Check universe appropriateness for user sets.
}
