% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lola.R
\name{checkUniverseAppropriateness}
\alias{checkUniverseAppropriateness}
\title{Check universe appropriateness}
\usage{
checkUniverseAppropriateness(userSets, userUniverse, cores = 1, fast = FALSE)
}
\arguments{
\item{userSets}{A RegionSet or list of RegionSets}

\item{userUniverse}{A RegionSet}

\item{cores}{Number of cores (currently unused)}

\item{fast}{Logical (currently unused)}
}
\value{
A data.frame with coverage diagnostics per user set
}
\description{
Check universe appropriateness
}
