% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-methods.R
\name{clusterRegions}
\alias{clusterRegions}
\alias{clusterRegions,RegionSet-method}
\alias{clusterRegions,ANY-method}
\title{Cluster regions via single-linkage with a stitching radius}
\usage{
clusterRegions(x, maxGap = 0L, ...)

\S4method{clusterRegions}{RegionSet}(x, maxGap = 0L, ...)

\S4method{clusterRegions}{ANY}(x, maxGap = 0L, ...)
}
\arguments{
\item{x}{A RegionSet, GRanges, file path, or data.frame}

\item{maxGap}{Maximum bp gap between regions to link (default 0)}

\item{...}{ignored}
}
\value{
Integer vector of 0-based cluster IDs in original region order
}
\description{
Assigns a cluster ID to each region. Regions within
\code{maxGap} bp of another region on the same chromosome are linked
into the same cluster; chromosome boundaries always break clusters.
}
