% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{compare_store}
\alias{compare_store}
\title{Compare two collections}
\usage{
compare_store(store_ptr, digest_a, digest_b)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{digest_a}{First collection digest}

\item{digest_b}{Second collection digest}
}
\description{
Compare two collections
}
