% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-methods.R
\name{concat}
\alias{concat}
\alias{concat,ANY-method}
\title{Combine two region sets without merging}
\usage{
concat(x, y)

\S4method{concat}{ANY}(x, y)
}
\arguments{
\item{x}{A RegionSet, GRanges, file path, or data.frame}

\item{y}{A RegionSet, GRanges, file path, or data.frame}
}
\value{
A RegionSet with all regions from both sets
}
\description{
Combine two region sets without merging
}
