% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-methods.R
\name{consensus}
\alias{consensus}
\title{Compute consensus regions from multiple region sets}
\usage{
consensus(sets)
}
\arguments{
\item{sets}{A list of RegionSet objects, GRanges objects, file paths,
or data.frames}
}
\value{
A data.frame with chr, start, end, count columns (0-based half-open)
}
\description{
Given a list of region sets, computes the union of all regions
and annotates each union region with the number of input sets that overlap
it.
}
