% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-methods.R
\name{countOverlaps,RegionSet,RegionSet-method}
\alias{countOverlaps,RegionSet,RegionSet-method}
\alias{countOverlaps,RegionSet,ANY-method}
\alias{countOverlaps,ANY,RegionSet-method}
\title{Count overlaps per query region}
\usage{
\S4method{countOverlaps}{RegionSet,RegionSet}(
  query,
  subject,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  ...
)

\S4method{countOverlaps}{RegionSet,ANY}(
  query,
  subject,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  ...
)

\S4method{countOverlaps}{ANY,RegionSet}(
  query,
  subject,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  ...
)
}
\arguments{
\item{query}{A RegionSet, GRanges, file path, or data.frame}

\item{subject}{A RegionSet, GRanges, file path, or data.frame}

\item{maxgap}{ignored}

\item{minoverlap}{Minimum overlap in base pairs (default 1)}

\item{type}{ignored}

\item{...}{ignored}
}
\value{
Integer vector of length equal to the number of query regions
}
\description{
For each query region, counts the number of subject regions
that overlap by at least \code{minoverlap} base pairs.
}
