% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refget.R
\name{digest_fasta}
\alias{digest_fasta}
\title{digest fasta}
\usage{
digest_fasta(fasta)
}
\arguments{
\item{fasta}{a filepath string to a fasta file}
}
\value{
a SequenceCollection object
}
\description{
digest a fasta file given a filepath
}
\examples{
\dontrun{
fasta_path <- '../../gtars/tests/data/fasta/base.fa'
result <- gtars::digest_fasta(fasta_path)
}

}
