% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{digest_fasta_raw}
\alias{digest_fasta_raw}
\title{Digest fasta file}
\usage{
digest_fasta_raw(fasta)
}
\arguments{
\item{fasta}{A filepath string to a fasta file.}
}
\description{
Digest fasta file
}
