% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{disable_ancillary_digests_store}
\alias{disable_ancillary_digests_store}
\title{Disable ancillary digest computation}
\usage{
disable_ancillary_digests_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
Disable ancillary digest computation
}
