% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{disable_attribute_index_store}
\alias{disable_attribute_index_store}
\title{Disable attribute index}
\usage{
disable_attribute_index_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
Disable attribute index
}
