% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{disable_encoding_store}
\alias{disable_encoding_store}
\title{Disable encoding, use raw byte storage}
\usage{
disable_encoding_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
Disable encoding, use raw byte storage
}
