% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{disable_persistence_store}
\alias{disable_persistence_store}
\title{Disable disk persistence}
\usage{
disable_persistence_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
Disable disk persistence
}
