% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-methods.R
\name{disjoin,RegionSet-method}
\alias{disjoin,RegionSet-method}
\alias{disjoin,ANY-method}
\title{Break regions into non-overlapping disjoint pieces}
\usage{
\S4method{disjoin}{RegionSet}(x, ...)

\S4method{disjoin}{ANY}(x, ...)
}
\arguments{
\item{x}{A RegionSet, GRanges, file path, or data.frame}

\item{...}{ignored}
}
\value{
A RegionSet with disjoint regions
}
\description{
Break regions into non-overlapping disjoint pieces
}
