% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-methods.R
\name{distribution}
\alias{distribution}
\alias{distribution,RegionSet-method}
\alias{distribution,ANY-method}
\title{Region distribution across bins}
\usage{
distribution(x, nBins = 250L, ...)

\S4method{distribution}{RegionSet}(x, nBins = 250L, ...)

\S4method{distribution}{ANY}(x, nBins = 250L, ...)
}
\arguments{
\item{x}{A RegionSet, GRanges, file path, or data.frame}

\item{nBins}{Number of bins (default 250)}

\item{...}{ignored}
}
\value{
A data.table compatible with plotChromBins
}
\description{
Partitions the genome into fixed-size windows and counts
region overlaps per bin. Output is compatible with
GenomicDistributions::plotChromBins.
}
