% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{enable_ancillary_digests_store}
\alias{enable_ancillary_digests_store}
\title{Enable ancillary digest computation}
\usage{
enable_ancillary_digests_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
Enable ancillary digest computation
}
