% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{enable_attribute_index_store}
\alias{enable_attribute_index_store}
\title{Enable attribute index}
\usage{
enable_attribute_index_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
Enable attribute index
}
