% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{enable_encoding_store}
\alias{enable_encoding_store}
\title{Enable 2-bit encoding for space efficiency}
\usage{
enable_encoding_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
Enable 2-bit encoding for space efficiency
}
