% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{enable_persistence_store}
\alias{enable_persistence_store}
\title{Enable disk persistence}
\usage{
enable_persistence_store(store_ptr, path)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{path}{Directory for storing sequences and metadata}
}
\description{
Enable disk persistence
}
