% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{find_collections_by_attribute_store}
\alias{find_collections_by_attribute_store}
\title{Find collections by attribute digest}
\usage{
find_collections_by_attribute_store(store_ptr, attr_name, attr_digest)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{attr_name}{Attribute name (names, lengths, sequences, etc.)}

\item{attr_digest}{The digest to search for}
}
\description{
Find collections by attribute digest
}
