% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-methods.R
\name{flank,RegionSet-method}
\alias{flank,RegionSet-method}
\alias{flank,ANY-method}
\title{Generate flanking regions}
\usage{
\S4method{flank}{RegionSet}(x, width, start = TRUE, both = FALSE, use.names = TRUE, ...)

\S4method{flank}{ANY}(x, width, start = TRUE, both = FALSE, use.names = TRUE, ...)
}
\arguments{
\item{x}{A RegionSet, GRanges, file path, or data.frame}

\item{width}{Flank width in base pairs}

\item{start}{If TRUE (default), flank upstream of start; if FALSE,
flank downstream of end}

\item{both}{If TRUE, flank on both sides}

\item{...}{ignored}
}
\value{
A RegionSet with flanking regions
}
\description{
Generate flanking regions
}
