% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-methods.R
\name{gaps,RegionSet-method}
\alias{gaps,RegionSet-method}
\alias{gaps,character-method}
\alias{gaps,data.frame-method}
\title{Return gaps between regions per chromosome}
\usage{
\S4method{gaps}{RegionSet}(x, start = NA, end = NA, ...)

\S4method{gaps}{character}(x, start = NA, end = NA, ...)

\S4method{gaps}{data.frame}(x, start = NA, end = NA, ...)
}
\arguments{
\item{x}{A RegionSet, GRanges, file path, or data.frame}

\item{...}{ignored}
}
\value{
A RegionSet with gap regions
}
\description{
Return gaps between regions per chromosome
}
