% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{gda_gene_model}
\alias{gda_gene_model}
\title{Extract the gene model from a GDA annotation pointer.}
\usage{
gda_gene_model(gda_ptr)
}
\arguments{
\item{gda_ptr}{External pointer to a GenomicDistAnnotation}
}
\description{
Extract the gene model from a GDA annotation pointer.
}
