% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{gda_tss_index}
\alias{gda_tss_index}
\title{Derive a TssIndex from a GDA annotation (using strand-aware TSS positions).}
\usage{
gda_tss_index(gda_ptr)
}
\arguments{
\item{gda_ptr}{External pointer to a GenomicDistAnnotation}
}
\description{
Derive a TssIndex from a GDA annotation (using strand-aware TSS positions).
}
