% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_aliases_for_collection_store}
\alias{get_aliases_for_collection_store}
\title{Get all aliases for a collection}
\usage{
get_aliases_for_collection_store(store_ptr, digest)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{digest}{Collection digest}
}
\description{
Get all aliases for a collection
}
