% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_attribute_store}
\alias{get_attribute_store}
\title{Get attribute array by digest}
\usage{
get_attribute_store(store_ptr, attr_name, attr_digest)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{attr_name}{Attribute name}

\item{attr_digest}{Attribute digest}
}
\description{
Get attribute array by digest
}
