% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_cache_path_store}
\alias{get_cache_path_store}
\title{Get cache path}
\usage{
get_cache_path_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
Get cache path
}
