% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_collection_level1_store}
\alias{get_collection_level1_store}
\title{Get level 1 representation (attribute digests) for a collection}
\usage{
get_collection_level1_store(store_ptr, digest)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{digest}{Collection digest}
}
\description{
Get level 1 representation (attribute digests) for a collection
}
