% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_collection_metadata_store}
\alias{get_collection_metadata_store}
\title{Get collection metadata}
\usage{
get_collection_metadata_store(store_ptr, digest)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{digest}{Collection digest}
}
\description{
Get collection metadata
}
