% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_collection_store}
\alias{get_collection_store}
\title{Get a collection with all sequences loaded}
\usage{
get_collection_store(store_ptr, digest)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{digest}{Collection digest}
}
\description{
Get a collection with all sequences loaded
}
