% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_fhr_metadata_store}
\alias{get_fhr_metadata_store}
\title{Get FHR metadata for a collection}
\usage{
get_fhr_metadata_store(store_ptr, collection_digest)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{collection_digest}{Collection digest}
}
\description{
Get FHR metadata for a collection
}
