% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_quiet_store}
\alias{get_quiet_store}
\title{Get quiet mode status}
\usage{
get_quiet_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
Get quiet mode status
}
