% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_remote_url_store}
\alias{get_remote_url_store}
\title{Get remote URL}
\usage{
get_remote_url_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
Get remote URL
}
