% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_seqs_bed_file_store}
\alias{get_seqs_bed_file_store}
\title{Extract sequences from store using BED file as FASTA}
\usage{
get_seqs_bed_file_store(
  store_ptr,
  collection_digest,
  bed_file_path,
  output_file_path
)
}
\arguments{
\item{store_ptr}{External pointer to GlobalRefgetStore}

\item{collection_digest}{Sequence collection digest}

\item{bed_file_path}{Path to BED file}

\item{output_file_path}{Path to output FASTA file}
}
\description{
Extract sequences from store using BED file as FASTA
}
