% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_seqs_bed_file_to_vec_store}
\alias{get_seqs_bed_file_to_vec_store}
\title{Extract sequences from store using BED file into memory}
\usage{
get_seqs_bed_file_to_vec_store(store_ptr, collection_digest, bed_file_path)
}
\arguments{
\item{store_ptr}{External pointer to GlobalRefgetStore}

\item{collection_digest}{Sequence collection digest}

\item{bed_file_path}{Path to BED file}
}
\description{
Extract sequences from store using BED file into memory
}
