% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_sequence_by_alias_store}
\alias{get_sequence_by_alias_store}
\title{Get sequence by alias}
\usage{
get_sequence_by_alias_store(store_ptr, namespace, alias)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{namespace}{Alias namespace}

\item{alias}{Alias name}
}
\description{
Get sequence by alias
}
