% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_sequence_by_id_store}
\alias{get_sequence_by_id_store}
\title{Get sequence by ID from store}
\usage{
get_sequence_by_id_store(store_ptr, digest)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{digest}{Sequence digest}
}
\description{
Get sequence by ID from store
}
