% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_sequence_by_name_store}
\alias{get_sequence_by_name_store}
\title{Get sequence by collection and name}
\usage{
get_sequence_by_name_store(store_ptr, collection_digest, sequence_name)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{collection_digest}{Sequence collection digest}

\item{sequence_name}{Sequence name}
}
\description{
Get sequence by collection and name
}
