% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_sequence_metadata_store}
\alias{get_sequence_metadata_store}
\title{Get sequence metadata (no sequence data)}
\usage{
get_sequence_metadata_store(store_ptr, digest)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{digest}{Sequence digest}
}
\description{
Get sequence metadata (no sequence data)
}
