% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_sequence_store}
\alias{get_sequence_store}
\title{Get sequence by digest from store (supports SHA512t24u or MD5)}
\usage{
get_sequence_store(store_ptr, digest)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{digest}{Sequence digest (SHA512t24u or MD5)}
}
\description{
Get sequence by digest from store (supports SHA512t24u or MD5)
}
