% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_storage_mode_store}
\alias{get_storage_mode_store}
\title{Get current storage mode}
\usage{
get_storage_mode_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
Get current storage mode
}
