% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_substring_store}
\alias{get_substring_store}
\title{Get substring from sequence}
\usage{
get_substring_store(store_ptr, seq_digest, start, end)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{seq_digest}{Sequence digest}

\item{start}{Start position}

\item{end}{End position}
}
\description{
Get substring from sequence
}
