% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refget.R
\name{global_refget_store}
\alias{global_refget_store}
\title{global refget store}
\usage{
global_refget_store(mode)
}
\arguments{
\item{mode}{either 'raw' or 'encoded'}
}
\value{
a GlobalRefgetStore object
}
\description{
create a pointer to a new global refget store rust object
}
\examples{
\dontrun{
r_md5_digest('ATCG')
}

}
