% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{global_refget_store_raw}
\alias{global_refget_store_raw}
\title{Create a GlobalRefgetStore}
\usage{
global_refget_store_raw(mode)
}
\arguments{
\item{mode}{Storage mode: "raw" or "encoded"}
}
\description{
Create a GlobalRefgetStore
}
