% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicdist.R
\name{gtars_concat}
\alias{gtars_concat}
\title{Combine two region sets without merging}
\usage{
gtars_concat(a, b)
}
\arguments{
\item{a}{A GRanges object, file path, or RegionSet pointer}

\item{b}{A GRanges object, file path, or RegionSet pointer}
}
\value{
A RegionSet pointer with all regions from both sets
}
\description{
Combine two region sets without merging
}
