% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicdist.R
\name{gtars_consensus}
\alias{gtars_consensus}
\title{Compute consensus regions from multiple region sets}
\usage{
gtars_consensus(sets)
}
\arguments{
\item{sets}{A list of GRanges objects, file paths, or RegionSet pointers}
}
\value{
A data.frame with chr, start, end, count columns (0-based half-open)
}
\description{
Given a list of region sets, computes the union of all regions
and annotates each union region with the number of input sets that overlap
it. Useful for finding regions present in k-of-n replicates.
}
