% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicdist.R
\name{gtars_pintersect}
\alias{gtars_pintersect}
\title{Pairwise intersection by position}
\usage{
gtars_pintersect(a, b)
}
\arguments{
\item{a}{A GRanges object, file path, or RegionSet pointer}

\item{b}{A GRanges object, file path, or RegionSet pointer}
}
\value{
A RegionSet pointer with pairwise intersections
}
\description{
Pairwise intersection by position
}
