% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicdist.R
\name{gtars_promoters}
\alias{gtars_promoters}
\title{Generate promoter regions}
\usage{
gtars_promoters(query, upstream = 2000L, downstream = 200L)
}
\arguments{
\item{query}{A GRanges object, file path, or RegionSet pointer}

\item{upstream}{Number of bases upstream of start (default 2000)}

\item{downstream}{Number of bases downstream of start (default 200)}
}
\value{
A RegionSet pointer with promoter regions
}
\description{
Generate promoter regions
}
