% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicdist.R
\name{gtars_reduce}
\alias{gtars_reduce}
\title{Merge overlapping intervals}
\usage{
gtars_reduce(query)
}
\arguments{
\item{query}{A GRanges object, file path, or RegionSet pointer}
}
\value{
A RegionSet pointer with merged regions
}
\description{
Merge overlapping intervals
}
