% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicdist.R
\name{gtars_setdiff}
\alias{gtars_setdiff}
\title{Set difference of two region sets}
\usage{
gtars_setdiff(a, b)
}
\arguments{
\item{a}{A GRanges object, file path, or RegionSet pointer}

\item{b}{A GRanges object, file path, or RegionSet pointer}
}
\value{
A RegionSet pointer with the difference (a minus b)
}
\description{
Set difference of two region sets
}
