% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicdist.R
\name{gtars_union}
\alias{gtars_union}
\title{Merge two region sets into a non-overlapping set}
\usage{
gtars_union(a, b)
}
\arguments{
\item{a}{A GRanges object, file path, or RegionSet pointer}

\item{b}{A GRanges object, file path, or RegionSet pointer}
}
\value{
A RegionSet pointer with merged regions
}
\description{
Merge two region sets into a non-overlapping set
}
