% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{has_ancillary_digests_store}
\alias{has_ancillary_digests_store}
\title{Check if ancillary digests are enabled}
\usage{
has_ancillary_digests_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
Check if ancillary digests are enabled
}
