% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{has_attribute_index_store}
\alias{has_attribute_index_store}
\title{Check if attribute index is enabled}
\usage{
has_attribute_index_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
Check if attribute index is enabled
}
