% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igd.R
\name{igd_search}
\alias{igd_search}
\title{Search IGD Database}
\usage{
igd_search(database_path, query_path)
}
\arguments{
\item{database_path}{path to .igd database}

\item{query_path}{path to .bed file}
}
\value{
dataframe of overlap hits
}
\description{
Searches an IGD database for region overlaps with an input BED file
}
\examples{
\dontrun{
# Search database with default name
igd_search("path/to/database", "path/to/query/file")
}

}
