% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{import_fasta_store}
\alias{import_fasta_store}
\title{Import FASTA file into store}
\usage{
import_fasta_store(store_ptr, file_path)
}
\arguments{
\item{store_ptr}{External pointer to GlobalRefgetStore}

\item{file_path}{Path to FASTA file}
}
\description{
Import FASTA file into store
}
