% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{is_collection_loaded_store}
\alias{is_collection_loaded_store}
\title{Check if collection is fully loaded}
\usage{
is_collection_loaded_store(store_ptr, digest)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{digest}{Collection digest}
}
\description{
Check if collection is fully loaded
}
