% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{is_persisting_store}
\alias{is_persisting_store}
\title{Check if store is persisting to disk}
\usage{
is_persisting_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
Check if store is persisting to disk
}
