% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{iter_collections_store}
\alias{iter_collections_store}
\title{Iterate over all collections with sequences loaded}
\usage{
iter_collections_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
Iterate over all collections with sequences loaded
}
