% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{iter_sequences_store}
\alias{iter_sequences_store}
\title{Iterate over all sequences with data loaded}
\usage{
iter_sequences_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
Iterate over all sequences with data loaded
}
