% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-methods.R
\name{jaccard}
\alias{jaccard}
\alias{jaccard,ANY-method}
\title{Nucleotide-level Jaccard similarity}
\usage{
jaccard(x, y)

\S4method{jaccard}{ANY}(x, y)
}
\arguments{
\item{x}{A RegionSet, GRanges, file path, or data.frame}

\item{y}{A RegionSet, GRanges, file path, or data.frame}
}
\value{
Numeric value between 0 and 1
}
\description{
Computes |intersection| / |union| in base pairs between two
region sets. Both sets are reduced (merged) before comparison.
}
