% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{list_collections_store}
\alias{list_collections_store}
\title{List all collections in the store}
\usage{
list_collections_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
List all collections in the store
}
