% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{list_fhr_metadata_store}
\alias{list_fhr_metadata_store}
\title{List all collection digests with FHR metadata}
\usage{
list_fhr_metadata_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
List all collection digests with FHR metadata
}
