% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{list_region_sets}
\alias{list_region_sets}
\title{List region set filenames in a database.}
\usage{
list_region_sets(db, collections)
}
\arguments{
\item{db}{ExternalPtr to RegionDB}

\item{collections}{Optional character vector filter}
}
\description{
List region set filenames in a database.
}
