% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{list_sequence_alias_namespaces_store}
\alias{list_sequence_alias_namespaces_store}
\title{List all sequence alias namespaces}
\usage{
list_sequence_alias_namespaces_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
List all sequence alias namespaces
}
