% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{list_sequence_aliases_store}
\alias{list_sequence_aliases_store}
\title{List all aliases in a sequence namespace}
\usage{
list_sequence_aliases_store(store_ptr, namespace)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{namespace}{Namespace name}
}
\description{
List all aliases in a sequence namespace
}
