% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{list_sequences_store}
\alias{list_sequences_store}
\title{List all sequences in the store}
\usage{
list_sequences_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
List all sequences in the store
}
