% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lola.R
\name{listRegionSets}
\alias{listRegionSets}
\title{List region set filenames in a database}
\usage{
listRegionSets(regionDB, collections = NULL)
}
\arguments{
\item{regionDB}{An external pointer to a RegionDB}

\item{collections}{Optional character vector filter}
}
\value{
Character vector of filenames
}
\description{
List region set filenames in a database
}
