% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refget.R
\name{load_all_sequences}
\alias{load_all_sequences}
\title{Download and cache every sequence in the store (flow 3, in bulk).}
\usage{
load_all_sequences(store)
}
\arguments{
\item{store}{A RefgetStore.}
}
\description{
Download and cache every sequence in the store (flow 3, in bulk).
}
