% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{load_all_sequences_store}
\alias{load_all_sequences_store}
\title{Download and cache every sequence in the store (flow 3, in bulk)}
\usage{
load_all_sequences_store(store_ptr)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
Download and cache every sequence in the store (flow 3, in bulk)
}
