% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{load_binary_genome_assembly}
\alias{load_binary_genome_assembly}
\title{Load a binary genome assembly from a .fab file}
\usage{
load_binary_genome_assembly(fab_path)
}
\arguments{
\item{fab_path}{Path to a .fab binary FASTA file (created by gtars prep --fasta)}
}
\description{
Load a binary genome assembly from a .fab file
}
