% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{load_fhr_metadata_store}
\alias{load_fhr_metadata_store}
\title{Load FHR metadata from JSON file}
\usage{
load_fhr_metadata_store(store_ptr, collection_digest, path)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{collection_digest}{Collection digest}

\item{path}{Path to JSON file}
}
\description{
Load FHR metadata from JSON file
}
