% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refget.R
\name{load_from_directory}
\alias{load_from_directory}
\title{load refget store from directory}
\usage{
load_from_directory(root_path)
}
\arguments{
\item{root_path}{path to refget store}
}
\value{
a GlobalRefgetStore object
}
\description{
load refget store from directory
}
\examples{
\dontrun{
r_md5_digest('ATCG')
}

}
