% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{load_from_directory_store}
\alias{load_from_directory_store}
\title{Load store from directory}
\usage{
load_from_directory_store(root_path)
}
\arguments{
\item{root_path}{Path to read store from}
}
\description{
Load store from directory
}
