% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{load_gda_bin}
\alias{load_gda_bin}
\title{Load a GDA binary file and return a GenomicDistAnnotation pointer.}
\usage{
load_gda_bin(path)
}
\arguments{
\item{path}{Path to a .bin GDA file (produced by \verb{gtars prep})}
}
\description{
The returned pointer gives access to the gene model, partition list,
and TSS index contained in the pre-compiled binary.
}
