% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{load_genome_assembly}
\alias{load_genome_assembly}
\title{Load a genome assembly from a FASTA file}
\usage{
load_genome_assembly(fasta_path)
}
\arguments{
\item{fasta_path}{Path to a FASTA file}
}
\description{
Load a genome assembly from a FASTA file
}
