% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{load_region_db}
\alias{load_region_db}
\title{Load a LOLA region database from a folder.}
\usage{
load_region_db(db_location, collections, limit)
}
\arguments{
\item{db_location}{Path to the LOLA database folder}

\item{collections}{Optional character vector of collection names to load}

\item{limit}{Optional integer limit on files per collection}
}
\description{
Load a LOLA region database from a folder.
}
