% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{load_regionset}
\alias{load_regionset}
\title{Load a RegionSet from a BED/narrowPeak/gzip file}
\usage{
load_regionset(bed_path)
}
\arguments{
\item{bed_path}{Path to a BED, narrowPeak, or gzipped BED file}
}
\description{
Load a RegionSet from a BED/narrowPeak/gzip file
}
