% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{load_sequence_aliases_store}
\alias{load_sequence_aliases_store}
\title{Load sequence aliases from TSV file}
\usage{
load_sequence_aliases_store(store_ptr, namespace, path)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{namespace}{Namespace to load into}

\item{path}{Path to TSV file}
}
\description{
Load sequence aliases from TSV file
}
