% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{load_sequence_store}
\alias{load_sequence_store}
\title{Download and cache a single sequence's whole byte data (flow 3)}
\usage{
load_sequence_store(store_ptr, digest)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{digest}{Sequence digest (optionally "SQ."-prefixed)}
}
\description{
Download and cache a single sequence's whole byte data (flow 3)
}
