% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{load_signal_matrix_bin}
\alias{load_signal_matrix_bin}
\title{Load a SignalMatrix from a packed binary file.}
\usage{
load_signal_matrix_bin(path)
}
\arguments{
\item{path}{Path to a .bin signal matrix file (produced by \verb{gtars prep})}
}
\description{
Load a SignalMatrix from a packed binary file.
}
