% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{load_signal_matrix_tsv}
\alias{load_signal_matrix_tsv}
\title{Load a SignalMatrix from a TSV file.}
\usage{
load_signal_matrix_tsv(path)
}
\arguments{
\item{path}{Path to a TSV signal matrix file}
}
\description{
Load a SignalMatrix from a TSV file.
}
