% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicdist.R
\name{loadGenomeAssembly}
\alias{loadGenomeAssembly}
\title{Load a genome assembly from FASTA}
\usage{
loadGenomeAssembly(fasta_path)
}
\arguments{
\item{fasta_path}{Path to a FASTA file}
}
\value{
An external pointer to a GenomeAssembly
}
\description{
Loads a FASTA file into memory as a GenomeAssembly pointer.
This replaces the BSgenome object used in the original GenomicDistributions.
}
