% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lola.R
\name{loadRegionDB}
\alias{loadRegionDB}
\title{Load a LOLA region database from a folder}
\usage{
loadRegionDB(dbLocation, useCache = TRUE, limit = NULL, collections = NULL)
}
\arguments{
\item{dbLocation}{Path to the LOLA database folder}

\item{useCache}{Logical, currently ignored (reserved for future caching)}

\item{limit}{Optional integer limit on files per collection}

\item{collections}{Optional character vector of collection names to load}
}
\value{
An external pointer to a RegionDB object
}
\description{
Load a LOLA region database from a folder
}
