% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lola.R
\name{loadRegionDBFromBeds}
\alias{loadRegionDBFromBeds}
\title{Load a region database from BED file paths}
\usage{
loadRegionDBFromBeds(bedFiles, filenames = NULL)
}
\arguments{
\item{bedFiles}{Character vector of paths to BED files}

\item{filenames}{Optional character vector of display names}
}
\value{
An external pointer to a RegionDB object
}
\description{
Load a region database from BED file paths
}
