% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-methods.R
\name{narrow,RegionSet-method}
\alias{narrow,RegionSet-method}
\alias{narrow,character-method}
\alias{narrow,data.frame-method}
\title{Narrow regions by specifying a relative sub-range}
\usage{
\S4method{narrow}{RegionSet}(x, start = NA, end = NA, width = NA, use.names = TRUE)

\S4method{narrow}{character}(x, start = NA, end = NA, width = NA, use.names = TRUE)

\S4method{narrow}{data.frame}(x, start = NA, end = NA, width = NA, use.names = TRUE)
}
\arguments{
\item{x}{A RegionSet, GRanges, file path, or data.frame}

\item{start}{1-based relative start within each region (NA to omit)}

\item{end}{1-based relative end within each region (NA to omit)}

\item{width}{Width of the sub-range (NA to omit)}

\item{use.names}{ignored}
}
\value{
A RegionSet with narrowed regions
}
\description{
Narrow regions by specifying a relative sub-range
}
