% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-methods.R
\name{nearestNeighbors}
\alias{nearestNeighbors}
\alias{nearestNeighbors,RegionSet-method}
\alias{nearestNeighbors,ANY-method}
\title{Nearest neighbor distances}
\usage{
nearestNeighbors(x, ...)

\S4method{nearestNeighbors}{RegionSet}(x, ...)

\S4method{nearestNeighbors}{ANY}(x, ...)
}
\arguments{
\item{x}{A RegionSet, GRanges, file path, or data.frame}

\item{...}{ignored}
}
\value{
Numeric vector of nearest neighbor distances
}
\description{
Computes the distance from each region to its nearest
neighbor.
}
