% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-methods.R
\name{neighborDistances}
\alias{neighborDistances}
\alias{neighborDistances,RegionSet-method}
\alias{neighborDistances,ANY-method}
\title{Neighbor distances}
\usage{
neighborDistances(x, ...)

\S4method{neighborDistances}{RegionSet}(x, ...)

\S4method{neighborDistances}{ANY}(x, ...)
}
\arguments{
\item{x}{A RegionSet, GRanges, file path, or data.frame}

\item{...}{ignored}
}
\value{
Numeric vector of distances
}
\description{
Computes distances between consecutive regions on each
chromosome.
}
