% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{on_disk_store}
\alias{on_disk_store}
\title{Create a disk-backed RefgetStore}
\usage{
on_disk_store(path)
}
\arguments{
\item{path}{Path for storing sequences and metadata}
}
\description{
Create a disk-backed RefgetStore
}
