% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{open_local_store}
\alias{open_local_store}
\title{Open store from local directory}
\usage{
open_local_store(root_path)
}
\arguments{
\item{root_path}{Path to read store from}
}
\description{
Open store from local directory
}
