% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{open_remote_store}
\alias{open_remote_store}
\title{Open a remote RefgetStore with local caching}
\usage{
open_remote_store(cache_path, remote_url)
}
\arguments{
\item{cache_path}{Local directory to cache downloaded metadata and sequences}

\item{remote_url}{Base URL of the remote refget store}
}
\description{
Open a remote RefgetStore with local caching
}
