% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSetList-class.R
\name{pairwise_jaccard}
\alias{pairwise_jaccard}
\alias{pairwise_jaccard,RegionSetList-method}
\title{Pairwise Jaccard similarity matrix}
\usage{
pairwise_jaccard(x)

\S4method{pairwise_jaccard}{RegionSetList}(x)
}
\arguments{
\item{x}{A RegionSetList}
}
\value{
A numeric matrix of Jaccard similarities
}
\description{
Computes nucleotide-level Jaccard similarity for all pairs of
region sets. Returns an N x N symmetric matrix with 1.0 on the diagonal.
}
