% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-methods.R
\name{pintersect}
\alias{pintersect}
\alias{pintersect,ANY-method}
\title{Pairwise intersection by position}
\usage{
pintersect(x, y, ...)

\S4method{pintersect}{ANY}(x, y, ...)
}
\arguments{
\item{x}{A RegionSet, GRanges, file path, or data.frame}

\item{y}{A RegionSet, GRanges, file path, or data.frame}
}
\value{
A RegionSet with pairwise intersections
}
\description{
Pairwise intersection by position
}
