% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-methods.R
\name{promoters,RegionSet-method}
\alias{promoters,RegionSet-method}
\alias{promoters,character-method}
\alias{promoters,data.frame-method}
\title{Generate promoter regions}
\usage{
\S4method{promoters}{RegionSet}(x, upstream = 2000, downstream = 200, use.names = TRUE, ...)

\S4method{promoters}{character}(x, upstream = 2000, downstream = 200, use.names = TRUE, ...)

\S4method{promoters}{data.frame}(x, upstream = 2000, downstream = 200, use.names = TRUE, ...)
}
\arguments{
\item{x}{A RegionSet, GRanges, file path, or data.frame}

\item{upstream}{Bases upstream of start (default 2000)}

\item{downstream}{Bases downstream of start (default 200)}

\item{...}{ignored}
}
\value{
A RegionSet with promoter regions
}
\description{
Generate promoter regions
}
