% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_calc_expected_partitions}
\alias{r_calc_expected_partitions}
\title{Calculate observed vs expected partition enrichment}
\usage{
r_calc_expected_partitions(
  rs_ptr,
  partition_ptr,
  chrom_names,
  chrom_sizes,
  bp_proportion
)
}
\arguments{
\item{rs_ptr}{External pointer to query RegionSet}

\item{partition_ptr}{External pointer to PartitionList}

\item{chrom_names}{Character vector of chromosome names}

\item{chrom_sizes}{Integer vector of chromosome sizes}

\item{bp_proportion}{If TRUE, count base pairs; if FALSE, count regions}
}
\description{
Calculate observed vs expected partition enrichment
}
