% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_calc_feature_distances}
\alias{r_calc_feature_distances}
\title{Calculate signed distance from each query region to nearest feature.
Returns NA for regions on chromosomes without features.}
\usage{
r_calc_feature_distances(query_ptr, features_ptr)
}
\arguments{
\item{query_ptr}{External pointer to query RegionSet}

\item{features_ptr}{External pointer to features RegionSet}
}
\description{
Calculate signed distance from each query region to nearest feature.
Returns NA for regions on chromosomes without features.
}
