% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_calc_gc_content}
\alias{r_calc_gc_content}
\title{Calculate GC content for each region}
\usage{
r_calc_gc_content(rs_ptr, assembly_ptr, ignore_unk_chroms)
}
\arguments{
\item{rs_ptr}{External pointer to a RegionSet}

\item{assembly_ptr}{External pointer to a GenomeAssembly}

\item{ignore_unk_chroms}{Skip regions on chromosomes not in the assembly}
}
\description{
Calculate GC content for each region
}
