% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_calc_nearest_neighbors}
\alias{r_calc_nearest_neighbors}
\title{Calculate distance from each region to its nearest neighbor}
\usage{
r_calc_nearest_neighbors(rs_ptr)
}
\arguments{
\item{rs_ptr}{External pointer to a RegionSet}
}
\description{
Calculate distance from each region to its nearest neighbor
}
