% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_calc_neighbor_distances}
\alias{r_calc_neighbor_distances}
\title{Calculate distances between consecutive regions on each chromosome}
\usage{
r_calc_neighbor_distances(rs_ptr)
}
\arguments{
\item{rs_ptr}{External pointer to a RegionSet}
}
\description{
Calculate distances between consecutive regions on each chromosome
}
