% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_calc_partitions}
\alias{r_calc_partitions}
\title{Assign regions to genomic partitions}
\usage{
r_calc_partitions(rs_ptr, partition_ptr, bp_proportion)
}
\arguments{
\item{rs_ptr}{External pointer to query RegionSet}

\item{partition_ptr}{External pointer to PartitionList}

\item{bp_proportion}{If TRUE, count base pairs; if FALSE, count regions}
}
\description{
Assign regions to genomic partitions
}
