% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_calc_summary_signal_from_matrix}
\alias{r_calc_summary_signal_from_matrix}
\title{Calculate summary signal from a SignalMatrix pointer.}
\usage{
r_calc_summary_signal_from_matrix(rs_ptr, sm_ptr)
}
\arguments{
\item{rs_ptr}{External pointer to query RegionSet}

\item{sm_ptr}{External pointer to a SignalMatrix}
}
\description{
Calculate summary signal from a SignalMatrix pointer.
}
