% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_calc_tss_distances}
\alias{r_calc_tss_distances}
\title{Calculate absolute distance from each query region to nearest feature midpoint.
Returns NA for regions on chromosomes without features.}
\usage{
r_calc_tss_distances(query_ptr, features_ptr)
}
\arguments{
\item{query_ptr}{External pointer to query RegionSet}

\item{features_ptr}{External pointer to features RegionSet}
}
\description{
Calculate absolute distance from each query region to nearest feature midpoint.
Returns NA for regions on chromosomes without features.
}
