% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_calc_widths}
\alias{r_calc_widths}
\title{Calculate region widths (end - start)}
\usage{
r_calc_widths(rs_ptr)
}
\arguments{
\item{rs_ptr}{External pointer to a RegionSet}
}
\description{
Calculate region widths (end - start)
}
