% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_chromosome_statistics}
\alias{r_chromosome_statistics}
\title{Calculate per-chromosome statistics}
\usage{
r_chromosome_statistics(rs_ptr)
}
\arguments{
\item{rs_ptr}{External pointer to a RegionSet}
}
\description{
Calculate per-chromosome statistics
}
