% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_consensus}
\alias{r_consensus}
\title{Compute consensus regions from a list of RegionSet pointers}
\usage{
r_consensus(rs_list)
}
\arguments{
\item{rs_list}{An R list of external pointers to RegionSets}
}
\description{
Compute consensus regions from a list of RegionSet pointers
}
