% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_count_overlaps}
\alias{r_count_overlaps}
\title{Count the number of subject regions overlapping each query region}
\usage{
r_count_overlaps(query_ptr, subject_ptr, minoverlap)
}
\arguments{
\item{query_ptr}{External pointer to query RegionSet}

\item{subject_ptr}{External pointer to subject RegionSet}

\item{minoverlap}{Minimum overlap in base pairs (default 1)}
}
\description{
Count the number of subject regions overlapping each query region
}
