% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_disjoin}
\alias{r_disjoin}
\title{Break regions into non-overlapping disjoint pieces}
\usage{
r_disjoin(rs_ptr)
}
\arguments{
\item{rs_ptr}{External pointer to a RegionSet}
}
\description{
Break regions into non-overlapping disjoint pieces
}
