% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_find_overlaps}
\alias{r_find_overlaps}
\title{Find all overlapping (queryHits, subjectHits) pairs between two RegionSets}
\usage{
r_find_overlaps(query_ptr, subject_ptr, minoverlap)
}
\arguments{
\item{query_ptr}{External pointer to query RegionSet}

\item{subject_ptr}{External pointer to subject RegionSet}

\item{minoverlap}{Minimum overlap in base pairs (default 1)}
}
\description{
Find all overlapping (queryHits, subjectHits) pairs between two RegionSets
}
