% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_flank}
\alias{r_flank}
\title{Generate flanking regions}
\usage{
r_flank(rs_ptr, width, use_start, both)
}
\arguments{
\item{rs_ptr}{External pointer to a RegionSet}

\item{width}{Flank width in base pairs}

\item{use_start}{If TRUE, flank upstream of start; if FALSE, downstream of end}

\item{both}{If TRUE, flank on both sides of the anchor}
}
\description{
Generate flanking regions
}
