% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_gaps}
\alias{r_gaps}
\title{Return gaps between regions per chromosome}
\usage{
r_gaps(rs_ptr)
}
\arguments{
\item{rs_ptr}{External pointer to a RegionSet}
}
\description{
Return gaps between regions per chromosome
}
