% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_intersect}
\alias{r_intersect}
\title{Range-level intersection of two region sets}
\usage{
r_intersect(rs_ptr_a, rs_ptr_b)
}
\arguments{
\item{rs_ptr_a}{External pointer to RegionSet A}

\item{rs_ptr_b}{External pointer to RegionSet B}
}
\description{
Range-level intersection of two region sets
}
