% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_jaccard}
\alias{r_jaccard}
\title{Nucleotide-level Jaccard similarity between two region sets}
\usage{
r_jaccard(rs_ptr_a, rs_ptr_b)
}
\arguments{
\item{rs_ptr_a}{External pointer to RegionSet A}

\item{rs_ptr_b}{External pointer to RegionSet B}
}
\description{
Nucleotide-level Jaccard similarity between two region sets
}
