% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_narrow}
\alias{r_narrow}
\title{Narrow regions by specifying a relative sub-range}
\usage{
r_narrow(rs_ptr, start, end, width)
}
\arguments{
\item{rs_ptr}{External pointer to a RegionSet}

\item{start}{1-based relative start (NA to omit)}

\item{end}{1-based relative end (NA to omit)}

\item{width}{Width of the sub-range (NA to omit)}
}
\description{
Narrow regions by specifying a relative sub-range
}
