% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_pairwise_jaccard}
\alias{r_pairwise_jaccard}
\title{Pairwise Jaccard similarity matrix for a RegionSetList.}
\usage{
r_pairwise_jaccard(rsl_ptr)
}
\arguments{
\item{rsl_ptr}{External pointer to a RegionSetList}
}
\description{
Pairwise Jaccard similarity matrix for a RegionSetList.
}
