% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_partition_list_from_gtf}
\alias{r_partition_list_from_gtf}
\title{Build a PartitionList from a GTF file}
\usage{
r_partition_list_from_gtf(
  gtf_path,
  filter_protein_coding,
  convert_ensembl_ucsc,
  core_prom,
  prox_prom,
  chrom_names,
  chrom_sizes_vec
)
}
\arguments{
\item{gtf_path}{Path to a GTF or GTF.gz file}

\item{filter_protein_coding}{Keep only protein-coding genes}

\item{convert_ensembl_ucsc}{Prepend "chr" to bare chromosome names}

\item{core_prom}{Core promoter size in bp}

\item{prox_prom}{Proximal promoter size in bp}

\item{chrom_names}{Chromosome names for trim (empty to skip)}

\item{chrom_sizes_vec}{Chromosome sizes for trim (parallel to chrom_names)}
}
\description{
Build a PartitionList from a GTF file
}
