% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_pintersect}
\alias{r_pintersect}
\title{Pairwise intersection by index position}
\usage{
r_pintersect(rs_ptr_a, rs_ptr_b)
}
\arguments{
\item{rs_ptr_a}{External pointer to RegionSet A}

\item{rs_ptr_b}{External pointer to RegionSet B}
}
\description{
Pairwise intersection by index position
}
