% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_promoters}
\alias{r_promoters}
\title{Generate promoter regions relative to each region's start}
\usage{
r_promoters(rs_ptr, upstream, downstream)
}
\arguments{
\item{rs_ptr}{External pointer to a RegionSet}

\item{upstream}{Bases upstream of start}

\item{downstream}{Bases downstream of start}
}
\description{
Generate promoter regions relative to each region's start
}
