% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_reduce}
\alias{r_reduce}
\title{Merge overlapping and adjacent intervals per chromosome}
\usage{
r_reduce(rs_ptr)
}
\arguments{
\item{rs_ptr}{External pointer to a RegionSet}
}
\description{
Merge overlapping and adjacent intervals per chromosome
}
