% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_region_distribution}
\alias{r_region_distribution}
\title{Calculate region distribution across chromosome bins}
\usage{
r_region_distribution(rs_ptr, n_bins)
}
\arguments{
\item{rs_ptr}{External pointer to a RegionSet}

\item{n_bins}{Number of bins (default 250)}
}
\description{
Calculate region distribution across chromosome bins
}
