% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_resize}
\alias{r_resize}
\title{Resize regions to a fixed width}
\usage{
r_resize(rs_ptr, width, fix)
}
\arguments{
\item{rs_ptr}{External pointer to a RegionSet}

\item{width}{New width in base pairs}

\item{fix}{Anchor point: "start", "end", or "center"}
}
\description{
Resize regions to a fixed width
}
