% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_setdiff}
\alias{r_setdiff}
\title{Set difference: remove portions of A that overlap with B}
\usage{
r_setdiff(rs_ptr_a, rs_ptr_b)
}
\arguments{
\item{rs_ptr_a}{External pointer to RegionSet A}

\item{rs_ptr_b}{External pointer to RegionSet B}
}
\description{
Set difference: remove portions of A that overlap with B
}
