% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refget.R
\name{r_sha512t24u_digest}
\alias{r_sha512t24u_digest}
\title{create sha512t24u_digest}
\usage{
r_sha512t24u_digest(readable)
}
\arguments{
\item{readable}{a readable string}
}
\value{
a sha512t24u digest on success
}
\description{
create a sha512t24u digest from a readable string
}
\examples{
\dontrun{
r_sha512t24u_digest('ATCG')
}

}
