% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_shift}
\alias{r_shift}
\title{Shift all regions by a fixed offset}
\usage{
r_shift(rs_ptr, offset)
}
\arguments{
\item{rs_ptr}{External pointer to a RegionSet}

\item{offset}{Integer offset (positive = downstream, negative = upstream)}
}
\description{
Shift all regions by a fixed offset
}
