% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_trim}
\alias{r_trim}
\title{Trim regions to chromosome boundaries}
\usage{
r_trim(rs_ptr, chrom_names, chrom_sizes)
}
\arguments{
\item{rs_ptr}{External pointer to a RegionSet}

\item{chrom_names}{Character vector of chromosome names}

\item{chrom_sizes}{Integer vector of chromosome sizes}
}
\description{
Trim regions to chromosome boundaries
}
