% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_union}
\alias{r_union}
\title{Merge two region sets into a minimal non-overlapping set}
\usage{
r_union(rs_ptr_a, rs_ptr_b)
}
\arguments{
\item{rs_ptr_a}{External pointer to RegionSet A}

\item{rs_ptr_b}{External pointer to RegionSet B}
}
\description{
Merge two region sets into a minimal non-overlapping set
}
