% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{redefine_user_sets}
\alias{redefine_user_sets}
\title{Redefine user sets in terms of universe regions.}
\usage{
redefine_user_sets(user_sets_list, universe_ptr)
}
\arguments{
\item{user_sets_list}{R list of RegionSet external pointers}

\item{universe_ptr}{ExternalPtr to universe RegionSet}
}
\description{
Redefine user sets in terms of universe regions.
}
