% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lola.R
\name{redefineUserSets}
\alias{redefineUserSets}
\title{Redefine user sets in terms of universe regions}
\usage{
redefineUserSets(userSets, userUniverse, cores = 1)
}
\arguments{
\item{userSets}{A RegionSet or list of RegionSets}

\item{userUniverse}{A RegionSet}

\item{cores}{Number of cores (currently unused)}
}
\value{
A list of RegionSet objects
}
\description{
Redefine user sets in terms of universe regions
}
