% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refget.R
\name{refget_store}
\alias{refget_store}
\title{global refget store}
\usage{
refget_store(mode)
}
\arguments{
\item{mode}{either 'raw' or 'encoded'}
}
\value{
a RefgetStore object
}
\description{
create a pointer to a new global refget store rust object
}
\examples{
\dontrun{
store <- refget_store('raw')
fasta_path <- '../../gtars/tests/data/fasta/base.fa'
import_fasta(store, fasta_path)
seq <- get_sequence_by_id(store, 'iYtREV555dUFKg2_agSJW6suquUyPpMw')
}

}
