% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refget.R
\name{refget_store_on_disk}
\alias{refget_store_on_disk}
\title{Create a disk-backed RefgetStore}
\usage{
refget_store_on_disk(path)
}
\arguments{
\item{path}{Directory path for storing sequences and metadata}
}
\value{
A RefgetStore object
}
\description{
Create a RefgetStore that persists to disk.
}
\examples{
\dontrun{
store <- refget_store_on_disk("/path/to/store")
}

}
