% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refget.R
\name{refget_store_open_local}
\alias{refget_store_open_local}
\title{Open a local RefgetStore}
\usage{
refget_store_open_local(path)
}
\arguments{
\item{path}{Path to the RefgetStore directory}
}
\value{
A RefgetStore object
}
\description{
Load an existing RefgetStore from a directory.
}
\examples{
\dontrun{
store <- refget_store_open_local("/path/to/store")
}

}
