% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refget.R
\name{refget_store_open_remote}
\alias{refget_store_open_remote}
\title{Open a remote RefgetStore}
\usage{
refget_store_open_remote(cache_path, remote_url)
}
\arguments{
\item{cache_path}{Local directory for caching downloaded data}

\item{remote_url}{URL of the remote RefgetStore}
}
\value{
A RefgetStore object
}
\description{
Open a remote RefgetStore with local caching.
}
\examples{
\dontrun{
store <- refget_store_open_remote("/local/cache", "https://example.com/store")
}

}
