% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{refget_store_raw}
\alias{refget_store_raw}
\title{Create an in-memory RefgetStore}
\usage{
refget_store_raw(mode)
}
\arguments{
\item{mode}{Storage mode: "raw" or "encoded"}
}
\description{
Create an in-memory RefgetStore
}
