% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{regiondb_anno}
\alias{regiondb_anno}
\title{Get per-file annotations from a RegionDB as a data.frame-like list.}
\usage{
regiondb_anno(db)
}
\arguments{
\item{db}{ExternalPtr to RegionDB}
}
\description{
Returns a list with columns: filename, cellType, description, tissue,
dataSource, antibody, treatment, collection. This matches the structure
of R LOLA's regionDB$regionAnno.
}
