% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{regiondb_collection_anno}
\alias{regiondb_collection_anno}
\title{Get collection-level annotations from a RegionDB as a data.frame-like list.}
\usage{
regiondb_collection_anno(db)
}
\arguments{
\item{db}{ExternalPtr to RegionDB}
}
\description{
Get collection-level annotations from a RegionDB as a data.frame-like list.
}
