% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{regiondb_region_set}
\alias{regiondb_region_set}
\title{Get a single RegionSet from a RegionDB by 1-based index.}
\usage{
regiondb_region_set(db, index)
}
\arguments{
\item{db}{ExternalPtr to RegionDB}

\item{index}{1-based integer index into the region sets}
}
\description{
Get a single RegionSet from a RegionDB by 1-based index.
}
