% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lola.R
\name{regionDBCollectionAnno}
\alias{regionDBCollectionAnno}
\title{Get collection-level annotations from a RegionDB}
\usage{
regionDBCollectionAnno(regionDB)
}
\arguments{
\item{regionDB}{An external pointer to a RegionDB (from loadRegionDB)}
}
\value{
A data.table with columns: collectionname, collector, date,
source, description
}
\description{
Get collection-level annotations from a RegionDB
}
