% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lola.R
\name{regionDBRegionSet}
\alias{regionDBRegionSet}
\title{Get a single RegionSet from a RegionDB by index}
\usage{
regionDBRegionSet(regionDB, index)
}
\arguments{
\item{regionDB}{An external pointer to a RegionDB (from loadRegionDB)}

\item{index}{Integer (1-based) index into the region sets}
}
\value{
A RegionSet object
}
\description{
Retrieves a single RegionSet from a loaded RegionDB, returning it
as a RegionSet S4 object that can be converted to GRanges via
\code{as_granges()}.
}
