% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{regionset_length}
\alias{regionset_length}
\title{Get the number of regions in a RegionSet}
\usage{
regionset_length(rs_ptr)
}
\arguments{
\item{rs_ptr}{External pointer to a RegionSet}
}
\description{
Get the number of regions in a RegionSet
}
