% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicdist.R
\name{regionset_to_df}
\alias{regionset_to_df}
\title{Convert RegionSet to data.frame}
\usage{
regionset_to_df(rs)
}
\arguments{
\item{rs}{An external pointer to a RegionSet}
}
\value{
A data.frame with chr, start, end columns
}
\description{
Converts an ExternalPtr\if{html}{\out{<RegionSet>}} to a data.frame with
chr, start, end columns in 0-based half-open (BED) coordinates.
}
