% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{regionset_to_vectors}
\alias{regionset_to_vectors}
\title{Extract chr/start/end vectors from a RegionSet pointer}
\usage{
regionset_to_vectors(rs_ptr)
}
\arguments{
\item{rs_ptr}{External pointer to a RegionSet}
}
\description{
Extract chr/start/end vectors from a RegionSet pointer
}
