% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSetList-class.R
\docType{class}
\name{RegionSetList-class}
\alias{RegionSetList-class}
\title{RegionSetList class}
\description{
S4 class representing a collection of RegionSets, backed by
a Rust RegionSetList via externalptr. The gtars equivalent of GRangesList.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{An externalptr to the Rust RegionSetList}
}}

