% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{regionsetlist_concat}
\alias{regionsetlist_concat}
\title{Flatten a RegionSetList into a single RegionSet (no merging).}
\usage{
regionsetlist_concat(rsl)
}
\arguments{
\item{rsl}{ExternalPtr to RegionSetList}
}
\description{
Flatten a RegionSetList into a single RegionSet (no merging).
}
