% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{regionsetlist_from_db}
\alias{regionsetlist_from_db}
\title{Extract region sets from a RegionDB by 1-based indices as a RegionSetList.}
\usage{
regionsetlist_from_db(db, indices)
}
\arguments{
\item{db}{ExternalPtr to RegionDB}

\item{indices}{Integer vector of 1-based indices}
}
\description{
Extract region sets from a RegionDB by 1-based indices as a RegionSetList.
}
