% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{regionsetlist_from_sets}
\alias{regionsetlist_from_sets}
\title{Create a RegionSetList from a list of RegionSet external pointers.}
\usage{
regionsetlist_from_sets(sets_list)
}
\arguments{
\item{sets_list}{R list of RegionSet external pointers}
}
\description{
Create a RegionSetList from a list of RegionSet external pointers.
}
