% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{regionsetlist_get}
\alias{regionsetlist_get}
\title{Get a single RegionSet from a RegionSetList by 1-based index.}
\usage{
regionsetlist_get(rsl, index)
}
\arguments{
\item{rsl}{ExternalPtr to RegionSetList}

\item{index}{1-based integer index}
}
\description{
Get a single RegionSet from a RegionSetList by 1-based index.
}
