% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{regionsetlist_length}
\alias{regionsetlist_length}
\title{Get the number of region sets in a RegionSetList.}
\usage{
regionsetlist_length(rsl)
}
\arguments{
\item{rsl}{ExternalPtr to RegionSetList}
}
\description{
Get the number of region sets in a RegionSetList.
}
