% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{regionsetlist_names}
\alias{regionsetlist_names}
\title{Get the names from a RegionSetList, or NULL if no names.}
\usage{
regionsetlist_names(rsl)
}
\arguments{
\item{rsl}{ExternalPtr to RegionSetList}
}
\description{
Get the names from a RegionSetList, or NULL if no names.
}
