% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{remove_collection_alias_store}
\alias{remove_collection_alias_store}
\title{Remove a collection alias}
\usage{
remove_collection_alias_store(store_ptr, namespace, alias)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{namespace}{Namespace name}

\item{alias}{Alias name}
}
\description{
Remove a collection alias
}
