% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{remove_fhr_metadata_store}
\alias{remove_fhr_metadata_store}
\title{Remove FHR metadata for a collection}
\usage{
remove_fhr_metadata_store(store_ptr, collection_digest)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{collection_digest}{Collection digest}
}
\description{
Remove FHR metadata for a collection
}
