% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{remove_sequence_alias_store}
\alias{remove_sequence_alias_store}
\title{Remove a sequence alias}
\usage{
remove_sequence_alias_store(store_ptr, namespace, alias)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{namespace}{Namespace name}

\item{alias}{Alias name}
}
\description{
Remove a sequence alias
}
