% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-methods.R
\name{resize,RegionSet-method}
\alias{resize,RegionSet-method}
\alias{resize,character-method}
\alias{resize,data.frame-method}
\title{Resize regions to a fixed width}
\usage{
\S4method{resize}{RegionSet}(x, width, fix = "start", use.names = TRUE, ...)

\S4method{resize}{character}(x, width, fix = "start", use.names = TRUE, ...)

\S4method{resize}{data.frame}(x, width, fix = "start", use.names = TRUE, ...)
}
\arguments{
\item{x}{A RegionSet, GRanges, file path, or data.frame}

\item{width}{New width in base pairs}

\item{fix}{Anchor point: "start" (default), "end", or "center"}

\item{...}{ignored}
}
\value{
A RegionSet with resized regions
}
\description{
Resize regions to a fixed width
}
