% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{run_lola}
\alias{run_lola}
\title{Run LOLA enrichment analysis.}
\usage{
run_lola(user_sets_list, universe_ptr, db_ptr, min_overlap, direction)
}
\arguments{
\item{user_sets_list}{R list of RegionSet external pointers}

\item{universe_ptr}{ExternalPtr to universe RegionSet}

\item{db_ptr}{ExternalPtr to RegionDB}

\item{min_overlap}{Minimum base-pair overlap (default 1)}

\item{direction}{"enrichment" or "depletion"}
}
\description{
Run LOLA enrichment analysis.
}
