% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lola.R
\name{runLOLA}
\alias{runLOLA}
\title{Run LOLA enrichment analysis}
\usage{
runLOLA(
  userSets,
  userUniverse,
  regionDB,
  minOverlap = 1,
  cores = 1,
  redefineUserSets = FALSE,
  direction = "enrichment"
)
}
\arguments{
\item{userSets}{A RegionSet object or list of RegionSet objects}

\item{userUniverse}{A RegionSet object representing the universe}

\item{regionDB}{An external pointer to a RegionDB (from loadRegionDB)}

\item{minOverlap}{Minimum base-pair overlap (default 1)}

\item{cores}{Number of cores (currently unused, reserved for future parallelism)}

\item{redefineUserSets}{Logical, whether to redefine user sets against universe}

\item{direction}{"enrichment" or "depletion"}
}
\value{
A data.frame with LOLA results
}
\description{
Run LOLA enrichment analysis
}
