% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/sc.R
\name{sc_feature_matrix_info}
\alias{sc_feature_matrix_info}
\title{Get FeatureMatrix dimensions and metadata}
\usage{
sc_feature_matrix_info(ptr)

sc_feature_matrix_info(ptr)
}
\arguments{
\item{ptr}{External pointer to a FeatureMatrix}
}
\value{
A named list with n_features, n_cells, nnz, feature_type
}
\description{
Get FeatureMatrix dimensions and metadata

Get FeatureMatrix info
}
