% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/sc.R
\name{sc_filter_genes}
\alias{sc_filter_genes}
\title{Filter genes appearing in fewer than min_cells cells}
\usage{
sc_filter_genes(ptr, min_cells = 3L)

sc_filter_genes(ptr, min_cells = 3L)
}
\arguments{
\item{ptr}{External pointer to a FeatureMatrix}

\item{min_cells}{Minimum number of cells a gene must appear in (default 3)}
}
\value{
External pointer to a new filtered FeatureMatrix
}
\description{
Filter genes appearing in fewer than min_cells cells

Filter genes by minimum cell count
}
