% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/sc.R
\name{sc_find_neighbors}
\alias{sc_find_neighbors}
\title{Build KNN + SNN graph from a PCA embedding matrix}
\usage{
sc_find_neighbors(embeddings, k = 20L, prune_snn = 1/15)

sc_find_neighbors(embeddings, k = 20L, prune_snn = 1/15)
}
\arguments{
\item{embeddings}{Matrix (cells x PCs) from sc_run_pca or sc_run_rna_pipeline}

\item{k}{Number of nearest neighbors (default 20)}

\item{prune_snn}{Minimum Jaccard index for SNN edges (default 1/15)}

\item{embedding_data}{Flat numeric vector (column-major) of cell embeddings}

\item{nrow}{Number of rows (cells)}

\item{ncol}{Number of columns (PCs)}
}
\value{
An external pointer to an SnnGraph
}
\description{
Build KNN + SNN graph from a PCA embedding matrix

Build nearest neighbor graph (KNN + SNN)
}
