% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/sc.R
\name{sc_find_variable_features}
\alias{sc_find_variable_features}
\title{Find highly variable features using variance-stabilizing transform}
\usage{
sc_find_variable_features(ptr, n_features = 2000L)

sc_find_variable_features(ptr, n_features = 2000L)
}
\arguments{
\item{ptr}{External pointer to a FeatureMatrix}

\item{n_features}{Number of variable features to select (default 2000)}
}
\value{
Character vector of variable feature names
}
\description{
Find highly variable features using variance-stabilizing transform

Find highly variable features
}
