% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/sc.R
\name{sc_read_10x}
\alias{sc_read_10x}
\title{Read a 10X Chromium directory (matrix.mtx.gz, features.tsv.gz, barcodes.tsv.gz)}
\usage{
sc_read_10x(dir)

sc_read_10x(dir)
}
\arguments{
\item{dir}{Path to a 10X directory containing matrix.mtx.gz,
features.tsv.gz, and barcodes.tsv.gz}
}
\value{
An external pointer to a FeatureMatrix
}
\description{
Read a 10X Chromium directory (matrix.mtx.gz, features.tsv.gz, barcodes.tsv.gz)

Read a 10X Chromium directory
}
