% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{set_fhr_metadata_store}
\alias{set_fhr_metadata_store}
\title{Set FHR metadata for a collection}
\usage{
set_fhr_metadata_store(store_ptr, collection_digest, metadata_json)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{collection_digest}{Collection digest}

\item{metadata_json}{JSON string with FHR metadata}
}
\description{
Set FHR metadata for a collection
}
