% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{set_quiet_store}
\alias{set_quiet_store}
\title{Set quiet mode}
\usage{
set_quiet_store(store_ptr, quiet)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{quiet}{Whether to suppress output}
}
\description{
Set quiet mode
}
