% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{sha512t24u_digest}
\alias{sha512t24u_digest}
\title{Create sha512t24u digest}
\usage{
sha512t24u_digest(readable)
}
\arguments{
\item{readable}{A readable string representing a sequence.}
}
\description{
Create sha512t24u digest
}
