% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-methods.R
\name{shift,RegionSet-method}
\alias{shift,RegionSet-method}
\alias{shift,ANY-method}
\title{Shift regions by a fixed offset}
\usage{
\S4method{shift}{RegionSet}(x, shift = 0L, use.names = TRUE)

\S4method{shift}{ANY}(x, shift = 0L, use.names = TRUE)
}
\arguments{
\item{x}{A RegionSet, GRanges, file path, or data.frame}

\item{shift}{Integer offset (positive = downstream, negative = upstream)}

\item{use.names}{ignored}
}
\value{
A RegionSet with shifted regions
}
\description{
Shift regions by a fixed offset
}
