% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{stream_sequence_store}
\alias{stream_sequence_store}
\title{Stream a (sub)sequence's decoded bases (flow 2)}
\usage{
stream_sequence_store(store_ptr, seq_digest, start, end)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{seq_digest}{Sequence digest (optionally "SQ."-prefixed)}

\item{start}{Start position (0-based, inclusive) or NULL}

\item{end}{End position (0-based, exclusive) or NULL}
}
\description{
Stream a (sub)sequence's decoded bases (flow 2)
}
