% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-methods.R
\name{trim,RegionSet-method}
\alias{trim,RegionSet-method}
\alias{trim,character-method}
\alias{trim,data.frame-method}
\title{Trim regions to chromosome boundaries}
\usage{
\S4method{trim}{RegionSet}(x, chromSizes, ...)

\S4method{trim}{character}(x, chromSizes, ...)

\S4method{trim}{data.frame}(x, chromSizes, ...)
}
\arguments{
\item{x}{A RegionSet, GRanges, file path, or data.frame}

\item{...}{Must include \code{chromSizes}: a named integer vector of
chromosome sizes}
}
\value{
A RegionSet with trimmed regions
}
\description{
Clips regions to fit within chromosome boundaries. Regions
on unknown chromosomes are dropped.
}
