% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-methods.R
\name{widths}
\alias{widths}
\alias{widths,RegionSet-method}
\alias{widths,ANY-method}
\title{Region widths}
\usage{
widths(x, ...)

\S4method{widths}{RegionSet}(x, ...)

\S4method{widths}{ANY}(x, ...)
}
\arguments{
\item{x}{A RegionSet, GRanges, file path, or data.frame}

\item{...}{ignored}
}
\value{
Numeric vector of widths
}
\description{
Calculates the width (end - start) of each region.
}
