% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{write_store}
\alias{write_store}
\title{Write store using configured paths}
\usage{
write_store(store)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}
}
\description{
Write store using configured paths
}
