% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{write_store_to_directory_store}
\alias{write_store_to_directory_store}
\title{Write store to directory}
\usage{
write_store_to_directory_store(store_ptr, root_path, seqdata_path_template)
}
\arguments{
\item{store_ptr}{External pointer to RefgetStore}

\item{root_path}{Path to write store}

\item{seqdata_path_template}{Path template name (optional, pass empty string for default)}
}
\description{
Write store to directory
}
