% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{write_tokens_to_gtok}
\alias{write_tokens_to_gtok}
\title{Write tokens to a gtok file}
\usage{
write_tokens_to_gtok(filename, tokens)
}
\arguments{
\item{filename}{A string representing the path to the gtok file.}

\item{tokens}{The tokens to write.}
}
\description{
Write tokens to a gtok file
}
