% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicdist.R
\name{as_granges}
\alias{as_granges}
\title{Convert RegionSet to GRanges}
\usage{
as_granges(rs)
}
\arguments{
\item{rs}{An external pointer to a RegionSet}
}
\value{
A GRanges object
}
\description{
Converts an ExternalPtr\if{html}{\out{<RegionSet>}} to a GRanges object.
}
\details{
Coordinates are converted from 0-based half-open (BED) to 1-based
closed (GRanges).
}
