% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicdist.R
\name{calcGCContent}
\alias{calcGCContent}
\title{Calculate GC content}
\usage{
calcGCContent(query, ref, ignoreUnkChroms = FALSE)
}
\arguments{
\item{query}{A GRanges object, file path, or RegionSet pointer}

\item{ref}{A GenomeAssembly pointer (from loadGenomeAssembly)}

\item{ignoreUnkChroms}{If TRUE, skip regions on chromosomes not in the
assembly (default FALSE)}
}
\value{
Numeric vector of GC content values (0 to 1)
}
\description{
Calculates GC content (proportion of G and C bases) for each
region. Port of GenomicDistributions::calcGCContent.
}
\details{
Unlike the original GenomicDistributions which accepts a BSgenome
object, this function requires a GenomeAssembly loaded from a FASTA file
via \code{\link{loadGenomeAssembly}}.
}
