% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicdist.R
\name{calcNeighborDist}
\alias{calcNeighborDist}
\title{Calculate neighbor distances}
\usage{
calcNeighborDist(query)
}
\arguments{
\item{query}{A GRanges object, file path, or RegionSet pointer}
}
\value{
Numeric vector of distances
}
\description{
Computes distances between consecutive regions on each chromosome.
Port of GenomicDistributions::calcNeighborDist.
}
