% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicdist.R
\name{calcWidth}
\alias{calcWidth}
\title{Calculate region widths}
\usage{
calcWidth(query)
}
\arguments{
\item{query}{A GRanges object, file path, or RegionSet pointer}
}
\value{
Numeric vector of widths
}
\description{
Calculates the width (end - start) of each region.
Port of GenomicDistributions::calcWidth.
}
