% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_calc_dinucl_freq}
\alias{r_calc_dinucl_freq}
\title{Calculate per-region dinucleotide frequencies (percentages)}
\usage{
r_calc_dinucl_freq(rs_ptr, assembly_ptr)
}
\arguments{
\item{rs_ptr}{External pointer to a RegionSet}

\item{assembly_ptr}{External pointer to a GenomeAssembly}
}
\description{
Calculate per-region dinucleotide frequencies (percentages)
}
