% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSet-class.R, R/RegionSet-methods.R
\docType{class}
\name{RegionSet-class}
\alias{RegionSet-class}
\alias{union,RegionSet,RegionSet-method}
\alias{union,RegionSet,ANY-method}
\alias{union,ANY,RegionSet-method}
\alias{intersect,RegionSet,RegionSet-method}
\alias{intersect,RegionSet,ANY-method}
\alias{intersect,ANY,RegionSet-method}
\alias{setdiff,RegionSet,RegionSet-method}
\alias{setdiff,RegionSet,ANY-method}
\alias{setdiff,ANY,RegionSet-method}
\title{RegionSet class}
\usage{
\S4method{union}{RegionSet,RegionSet}(x, y)

\S4method{union}{RegionSet,ANY}(x, y)

\S4method{union}{ANY,RegionSet}(x, y)

\S4method{intersect}{RegionSet,RegionSet}(x, y)

\S4method{intersect}{RegionSet,ANY}(x, y)

\S4method{intersect}{ANY,RegionSet}(x, y)

\S4method{setdiff}{RegionSet,RegionSet}(x, y)

\S4method{setdiff}{RegionSet,ANY}(x, y)

\S4method{setdiff}{ANY,RegionSet}(x, y)
}
\description{
S4 class representing a set of genomic regions, backed by
a Rust RegionSet via externalptr. Carries an R-side strand vector.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{An externalptr to the Rust RegionSet}

\item{\code{strand}}{A character vector of strand values ("+", "-", or "*")}
}}

