% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/genomicdist.R
\name{regionset_length}
\alias{regionset_length}
\title{Get the number of regions in a RegionSet}
\usage{
regionset_length(rs)

regionset_length(rs)
}
\arguments{
\item{rs}{An external pointer to a RegionSet}

\item{rs_ptr}{External pointer to a RegionSet}
}
\value{
Integer count of regions
}
\description{
Get the number of regions in a RegionSet

Get number of regions in a RegionSet
}
